Refactor ingest (#1)
New modular ingest framework.

* Adds fixtures for example SIPs

* Adds minitar and nokogiri dependencies

* Adds some samples for testing

* Big refactor.

This is a big squash of several commits that were made during an experimental
phase. Since lots of the earlier commits were obsolted by the latter ones, and
since this gem is not in production anywhere, and still very young, I decided to
squash everything, and put in a pretty big commit with a lot of meat as to how
the ingest process is intended to work.

Instead of detailing the incremental steps of how I got to this point, I'll put
forth the main idea:

* The SIP object is a model for reading files on disk that are to be ingested.

* The Ingester object has two responsibilities:
  1. Read in the mapping configuration (a YAML file).
  1. Initiate the ingestion process

* The mapping file (once parsed) specifies one or more map operations.

* Each map operation consists of a fetch operation, followed by an assignment
  operation.

* MapOperation, Fetcher, and Assigner classes are used for modeling map, fetch,
  and assignment operations, respectively.

* The rules by which values are fetched from the SIP, and then assigned to
  another object, are specified in the mapping configuration.

That's the basic gist of how the ingest process is designed. The state of this
commit is work-in-progress, but it's time to get said progress synced to Github,
so that's what we're doing here.

* Allows SIPs of a single file

* Fixes tests for SIP class

* Adds spec for fetcher factory

* Adds class method Fetcher::Base.validate_options

This method is part of the interface for fetcher objects. For each extended
class, it should raise a Hyrax::Ingest::Errors::InvalidFetcherOptions error
unless the options are valid.

* Refactors fetcher and assigner factories.

Changes the factories to interpret options more literally when using them to
determine which class to instantiate. Before it was was less deterministic
and relying on a convention that wasn't totally clear, and would have required
additional edge-case logic.

* Adds tests for Hyrax::Ingest::Fetcher::Base

* More WIP on ingest refactor

* Changes Hyrax::Ingest::Fetcher::FromXML to Hyrax::Ingest::Fetcher::XMLFile.
* Fetcher factory determines which fetcher class to use by using fetch option as
  the class name, rather than inferring the class name from a combination of
  fetch options.
* Adds some error handling to ensure that fetcher class has a valid SIP object.

* Adds tests for assigning to AF model

* Allows reusing ActiveFedor model instances

When assigning to an ActiveFedora model using the ActiveFedorModel assigner, this change
stores a hash of instances keyed by an instance name (supplied by the user) to allow
for reusing the same instances when making assignments.

* Adds test for MapOperation#map

* Adds tests for Mapping object

* Fix for ActiveFedora environment on Travis