Add persistence (#3)
* Fixes example mapping config

Changing the purpose of this file to serve as useful inline documentation on
what the availalbe options are for the ingest configuration.

* Initials for persistance

* Privatize methods on Ingester class

Methods should be in private scope if they don't need to be a part of the public
interface.

* Implements persistence for ActiveFedora models

This implements the #save! method for Hyrax::Ingest::Assigner::ActiveFedoraModel,
which simply calls the #save! method on the ActiveFedora model.