Adds Reporting module
This is a refactor of the Logging module. Instead of classes including a Logging
module, and using the Logger class's public interface directly, there is now a
Reporter class that has a Logger class instance. This allows us to accumulate
summaries of logging output at the same time that we are writing to the log.